only send window size the first time it is noted in out of band data

+++ usr.bin/rlogin/rlogin.c
@@ -11,7 +11,7 @@ char copyright[] =
 #endif not lint
 
 #ifndef lint
-static char sccsid[] = "@(#)rlogin.c	5.5 (Berkeley) 09/12/85";
+static char sccsid[] = "@(#)rlogin.c	5.6 (Berkeley) 09/18/85";
 #endif not lint
 
 /*
@@ -222,8 +222,9 @@ done()
 writeroob()
 {
 
+	if (dosigwinch == 0)
+		sendwindow();
 	dosigwinch = 1;
-	sendwindow();
 }
 
 catchild()
@@ -369,6 +370,7 @@ oob()
 	int out = 1+1, atmark;
 	char waste[BUFSIZ], mark;
 	struct sgttyb sb;
+	static int didnotify = 0;
 
 	ioctl(1, TIOCFLUSH, (char *)&out);
 	for (;;) {
@@ -384,11 +386,12 @@ oob()
 			break;
 	}
 	recv(rem, &mark, 1, MSG_OOB);
-	if (mark & TIOCPKT_WINDOW) {
+	if (didnotify == 0 && (mark & TIOCPKT_WINDOW)) {
 		/*
 		 * Let server know about window size changes
 		 */
 		kill(getppid(), SIGURG);
+		didnotify = 1;
 	}
 	if (eight)
 		return;
